// app/api/users/route.ts
import User from "@/databases/user.model";
import { APIErrorResponse } from "@/global";
import handelError from "@/lib/handler/error";
import { ValidationError } from "@/lib/http-error";
import dbConnect from "@/lib/mongoose";
import { UserSchema } from "@/lib/validations";
import { NextResponse } from "next/server";

export async function GET() {
  try {
    await dbConnect();
    const users = await User.find();

    return NextResponse.json({ success: true, data: users }, { status: 200 });
  } catch (error) {
    return handelError(error, "api") as APIErrorResponse;
  }
}

export async function POST(request: Request) {
  try {
    await dbConnect();
    const body = await request.json();
    const validatedData = UserSchema.safeParse(body);

    if (!validatedData.success) {
      throw new ValidationError(validatedData.error.flatten().fieldErrors);
    }

    const { email, username } = validatedData.data;
    const existingUser = await User.findOne({ email });
    if (existingUser) {
      throw new Error("User already exist.");
    }

    const existingUsername = await User.findOne({ username });
    if (existingUsername) throw new Error("User already exist.");
    const newUser = await User.create(validatedData.data);
    return NextResponse.json({ success: true, data: newUser }, { status: 200 });
  } catch (error) {
    return handelError(error, "api") as APIErrorResponse;
  }
}
